+++ client/src/components/Chat/Menus/index.ts
@@ -1,3 +1,3 @@
 export { default as EndpointsMenu } from './EndpointsMenu';
 export { default as PresetsMenu } from './PresetsMenu';
-export { default as NewChat } from './NewChat';
+export { default as HeaderNewChat } from './HeaderNewChat';
